+++ apps/client/src/app/components/portfolio-summary/portfolio-summary.component.ts
@@ -2,7 +2,9 @@ import { NotificationService } from '@ghostfolio/client/core/notification/notifi
 import { getDateFnsLocale, getLocale } from '@ghostfolio/common/helper';
 import { PortfolioSummary, User } from '@ghostfolio/common/interfaces';
 import { translate } from '@ghostfolio/ui/i18n';
+import { GfValueComponent } from '@ghostfolio/ui/value';
 
+import { CommonModule } from '@angular/common';
 import {
   ChangeDetectionStrategy,
   Component,
@@ -11,6 +13,8 @@ import {
   OnChanges,
   Output
 } from '@angular/core';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { IonIcon } from '@ionic/angular/standalone';
 import { formatDistanceToNow } from 'date-fns';
 import { addIcons } from 'ionicons';
 import {
@@ -19,13 +23,13 @@ import {
 } from 'ionicons/icons';
 
 @Component({
-  selector: 'gf-portfolio-summary',
   changeDetection: ChangeDetectionStrategy.OnPush,
-  templateUrl: './portfolio-summary.component.html',
+  imports: [CommonModule, GfValueComponent, IonIcon, MatTooltipModule],
+  selector: 'gf-portfolio-summary',
   styleUrls: ['./portfolio-summary.component.scss'],
-  standalone: false
+  templateUrl: './portfolio-summary.component.html'
 })
-export class PortfolioSummaryComponent implements OnChanges {
+export class GfPortfolioSummaryComponent implements OnChanges {
   @Input() baseCurrency: string;
   @Input() hasPermissionToUpdateUserSettings: boolean;
   @Input() isLoading: boolean;